t13: NullPointerException  => kommt dadurch das es manachmal nicht klappt die files zulöschen, weil sie nocht nicht wieder freigegeben sind, ist atm so "gut" es geht gefixed NegativeArraySizeException => dadurch das es beim mal 2 rechnen der folgenlänge irgendwie nen overflow gibt trotz long ( 1,3milliarden *2) fix noch offen BufferUnderFlowException => durch das datei ende, dieser  fehler sollte behoben sein

+++ trunk/ad-praktikum-t13-t14/src/aufgabe2/data/FolgenWriter.java
@@ -32,7 +32,7 @@ public class FolgenWriter {
     }
     private void initBuffer(){
         // buffer der maximalen groeße erstellen
-        byteBuffer = ByteBuffer.allocate((Integer.SIZE / Byte.SIZE) * Writer.INTEGER_COUNT_PER_WRITE);
+        byteBuffer = ByteBuffer.allocate((Integer.SIZE / Byte.SIZE) * (int)Writer.INTEGER_COUNT_PER_WRITE);
         // bytebuffer als intbuffer interpretieren
         intBuffer = byteBuffer.asIntBuffer();
     }
